fix define test
